refactor(models): extract model catalog and display utilities
